requirements.txt
pywin32
ollama
requests
python-docx